+++ auth/src/main/java/com/owlexpress/auth/application/AuthServiceImpl.java
@@ -4,7 +4,7 @@ import com.owlexpress.auth.application.dto.request.UserLoginRequestDto;
 import com.owlexpress.auth.application.dto.response.UserInfoResponseDto;
 import com.owlexpress.auth.common.TokenProvider;
 import com.owlexpress.auth.infrastructure.UserClient;
-import com.owlexpress.auth.presentation.dto.CommonDto;
+import com.owlexpress.auth.presentation.AuthService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
